Support loading the backup descriptions file
The backup descriptions file contains information about each defined
backup, informing the script how to perform the backup, and what it
should include.  So far the data is only loaded, next steps are to
provide suitable access to the information and likely, create new object
instaces that will represent a specific backup job.