The restic_agent now handles "cleanup" action
The restic_agent will now correctly handle the "cleanup" action by
invoking restic to perform a cleanup of the backup and only keep data
from the last "n" backups.

The "KEEP_FULL" property of the backup description will instruct the
agent on how many backups to retain and will default to 3 if not
provided.