Refactor handler to receive all backup descriptions
This commit refactors the relationship between backup descriptions and
the handler.  Previously, the handler only received a single backup
description to process whereas it now received the full lost of backup
descriptions before being instructed which specific description to
process.

This change is necessary to allow some agents to establish when a backup
description has not been backed up and remains exposed as a risk.  An
example of such an agent is the gitolite_agent that can handle a backup
of all "exposed" repositories that are not included in any other backup
description.  In order to support this, it is necessary for it to be
aware of all backup descriptions.