Included a "gitolite_direct_agent" supporting a "backup" action
This commit introduces a minimal agent to handle backups from a gitolite
server that will clone each target repository directly onto the backup
medium.  As such it only supports the "backup" action.

This agent is useful when creating a backup of information that will be
needed in an unencrypted form from the backup medium.  As such, since
all data will be clearly readable, it should only be used for
non-sensitive data.