Ensure bfe operates normally when invoked from another directory
Currently when bfe is invoked from a directory other than where the
script is stored, it does not correctly resolve the path to the script.
Consequently, it is unable to include the additional script files
containing essential functionality.  It also doesn't resolve the
default location for configuration files.

This commit ensures that the script can determine where the it is
located, and how to find all the constituent parts.