Resolve global variable name clash
Owing to the way the descriptions type is introduced into bfe, via
pseudo OO, where a backup description instance is created with the name
"description" this will cause a global variable named "description_name"
to be generated.  This name is commonly used, and conflicts with
functionality in the "descriptions" type.  To resolve the conflit the
variable used in the "descriptions" type has been renamed accordingly.