Support backup status email notifications
The current script has no provision for automatic notification of backup
status.  This commit includes support for an email notificationt to be
sent on completion indicating the success or failure status.

It achieves this by using the mailx command and has been tested tested
with gmail servers to successfully send and email.  Some configuration
is required and details of this can be found in the comments of the
bfe.system.utils.sendEmail() method.