Include a git backup agent to work from url's
The bfe backup agents currently support backup of git repositories
served by gitolite.  This commit expands the capabilities of bfe by
including support for backup of git repositories from a complete url.
In doing so, it is possible to backup any repository that can be
described by a url, such as github repositories.