Renamed backup agents to follow "what" then "how" convention
The current naming of backup agents can be tricky to determine exactly
what is being backed up or how.  To clarify this, and make it instantly
understandable, all backup agents have been renamed so that they follow
the convention of "[what]_[how]".

When scanning the code, or more importantly, the backup descriptions it
is now much more understandable to know what will occur.