Added git_direct_agent
This commit introduces a new backup agent to handle the backup of git
repositories from url's directly to the backup medium.  ie. not using a
backup tool suh as restic, but instead clone the repository and copy all
files into the backup medium.  As such, since no encryption occurs, this
would not be suitable for sensitive information.

All copied data is verifyied via the use of file hashes to ensure data
consistency between the staged data and the restored files.