Ensure the correct exposed gitolite repositories are selected
To determine which gitolite repositories are left at risk and "exposed"
bfe reads all currently defined backup descriptions taking note of all
the repositories included in a backup.  Any that are not, are considered
at risk.
Part of the algorithm used first selects all the backup descriptions
of type "gitolite".  As the backup types have recently been changed this
test no longer works causing a regression in the backup of any exposed
repositories.
This commit repairs the regression by testing the backup type to see if
it starts with "gitolite" and if so, includes it in the processing.